48 - refactor detail screen
