+++ app/src/main/kotlin/com/architects/PokeArchApplication.kt
@@ -4,4 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class App: Application()
+class PokeArchApplication: Application()
