+++ app/src/main/kotlin/com/architects/pokearch/ui/components/placeHolders/NoSearchResultPlaceHolder.kt
@@ -1,0 +1,39 @@
+package com.architects.pokearch.ui.components.placeHolders
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.architects.pokearch.R
+
+@Composable
+fun NoSearchResultPlaceHolder(modifier: Modifier = Modifier) {
+    Column(
+        modifier = modifier.padding(16.dp),
+        verticalArrangement = Arrangement.Top,
+        horizontalAlignment = Alignment.CenterHorizontally,
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.not_found_placeholder),
+            contentScale = ContentScale.Crop,
+            contentDescription = null
+        )
+        Spacer(modifier = Modifier.padding(16.dp))
+        Text(
+            text = stringResource(id = R.string.no_search_results_text),
+            style = MaterialTheme.typography.headlineSmall,
+            textAlign = TextAlign.Center
+        )
+    }
+}
