+++ app/src/main/kotlin/com/architects/pokearch/ui/features/home/ui/HomeScreen.kt
@@ -1,0 +1,89 @@
+package com.architects.pokearch.ui.features.home.ui
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.architects.pokearch.ui.components.placeHolders.NoSearchResultPlaceHolder
+import com.architects.pokearch.ui.components.progressIndicators.ArchLoadingIndicator
+import com.architects.pokearch.ui.features.home.state.HomeUiState
+import com.architects.pokearch.ui.features.home.viewModel.HomeViewModel
+import kotlinx.coroutines.launch
+
+@Composable
+fun HomeScreen(
+    pokemonName: String,
+    onNavigationClick: (Int) -> Unit,
+    modifier: Modifier = Modifier,
+    viewModel: HomeViewModel = hiltViewModel(),
+) {
+
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val scope = rememberCoroutineScope()
+
+    LaunchedEffect(key1 = pokemonName) {
+        viewModel.getPokemonList(pokemonName)
+    }
+
+    Container(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(8.dp)
+    ) {
+
+        when (val state = uiState) {
+            is HomeUiState.Loading -> {
+                ArchLoadingIndicator()
+            }
+
+            is HomeUiState.Success -> {
+                HomeSuccessScreen(
+                    state = state,
+                    onItemClick = { pokemonId ->
+                        onNavigationClick(pokemonId)
+                    },
+                    onLoadMore = {
+                        scope.launch {
+                            viewModel.onLoadMore()
+                        }
+                    }
+                )
+            }
+
+            is HomeUiState.NoSearchResult -> {
+                NoSearchResultPlaceHolder(
+                    modifier = Modifier.align(Alignment.TopCenter)
+                )
+            }
+
+            is HomeUiState.Error -> {
+                Text(
+                    text = "Something went wrong"
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun Container(
+    modifier: Modifier = Modifier,
+    content: @Composable BoxScope.() -> Unit
+) {
+    Box(
+        modifier = modifier,
+        propagateMinConstraints = true,
+    ) {
+        content()
+    }
+}
