+++ app/src/main/kotlin/com/architects/pokearch/ui/features/home/ui/HomeSuccessScreen.kt
@@ -1,0 +1,66 @@
+package com.architects.pokearch.ui.features.home.ui
+
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyGridState
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.res.dimensionResource
+import com.architects.pokearch.R
+import com.architects.pokearch.ui.features.home.state.HomeUiState
+import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.filter
+
+@Composable
+fun HomeSuccessScreen(
+    state: HomeUiState.Success,
+    onItemClick: (Int) -> Unit,
+    onLoadMore: () -> Unit,
+) {
+    val lazyGridState = rememberLazyGridState()
+
+    LazyVerticalGrid(
+        state = lazyGridState,
+        columns = GridCells.Adaptive(dimensionResource(id = R.dimen.grid_cell_min_size))
+    ) {
+        items(state.pokemonList) {
+            HomeItem(
+                pokemon = it,
+                onItemClick = { pokemonId ->
+                    onItemClick(pokemonId)
+                }
+            )
+        }
+    }
+
+    lazyGridState.OnBottomReached {
+        onLoadMore()
+    }
+}
+
+
+@Composable
+fun LazyGridState.OnBottomReached(
+    onLoadMore: () -> Unit,
+) {
+    val shouldLoadMore = remember {
+        derivedStateOf {
+            val lastVisibleItem = layoutInfo.visibleItemsInfo.lastOrNull()
+                ?: return@derivedStateOf false
+
+            lastVisibleItem.index == layoutInfo.totalItemsCount - 1
+        }
+    }
+
+    LaunchedEffect(shouldLoadMore) {
+        snapshotFlow { shouldLoadMore.value }
+            .distinctUntilChanged()
+            .filter { it }
+            .collect { onLoadMore() }
+    }
+}
