+++ app/src/main/kotlin/com/architects/pokearch/ui/features/shakeNCatch/ui/ShakeNCatchScreen.kt
@@ -1,4 +1,4 @@
-package com.architects.pokearch.ui.feature
+package com.architects.pokearch.ui.features.shakeNCatch.ui
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -10,11 +10,12 @@ import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.architects.pokearch.ui.components.progressIndicators.ArchLoadingIndicator
+import com.architects.pokearch.ui.features.shakeNCatch.viewModel.ShakeNCatchViewModel
 
 @Composable
-fun RandomCatchScreen(
+fun ShakeNCatchScreen(
     modifier: Modifier = Modifier,
-    viewModel: FeatureViewModel = hiltViewModel(),
+    viewModel: ShakeNCatchViewModel = hiltViewModel(),
     onNavigationClick: (Int) -> Unit = {}
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
