+++ app/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
     <string name="app_name">PokeArch</string>
     <string name="home_title">Home</string>
     <string name="team_title">Team</string>
-    <string name="feature_title">Feature</string>
+    <string name="feature_title">Shake\'n\'Catch</string>
     <string name="error_async_image_painter">Error</string>
     <string name="base_stats">Base stats</string>
     <string name="search_label_place_holder">Search…</string>
